Return Logic to Passing
Reworks the existing logic path that leverages the new request code to
return the tests to a passing state. This doesn't guarantee that there
are no bugs in the new Change Request logic and is still transitional,
but the existing tests are now green. This leaves cleaning up old
queries and ensuring test queries are properly cleaned up undone as of
this commit.